Ensure proper alignment for SIMD operations
Because unaligned access is slow for AVX2 and crashes for SSE4.2. Now
the pixels for tiles and tile buffers in the paint engine get allocated
with 32 byte alignment, which is enough to cover both. Other operations
fall back to the sequential operations, but they're not nearly as hot,
so that shouldn't cause much of a difference.